refactor(profiler)!: name the plugin for the capability, not the domain
The unit being profiled is the fileset, not the dataset. Model source files live in
filesets too, and profiling them is the expected next kind -- so `nemo-datasets` was
asserting a scope the feature does not have, and would have had to be renamed the moment
a second kind landed.

  plugins/nemo-datasets            -> plugins/nemo-profiler
  nemo_datasets_plugin             -> nemo_profiler_plugin
  nemo_datasets_plugin.profiler    -> nemo_profiler_plugin.dataset
  ...profiler.file_source          -> nemo_profiler_plugin.source
  files/dataset_profile.py         -> files/profile.py

`file_source.py` moves up out of the engine because it is the one piece the kinds share:
`FileSource` is `list_files()` + `open()` and knows nothing about formats, so a model
profiler reads through the same seam. What the kinds do not share is the analysis or its
dependencies -- the dataset engine needs pyarrow, a model engine will need the checkpoint
libraries -- which is why each gets a subpackage here rather than a package of its own.
Splitting them into sibling plugins stays available if the engines turn out to share
nothing but the contract.

The contract keeps its home in `nemo_platform_plugin.files`: it is pydantic-only, and the
Files service owns the profile as a resource. Only its filename was over-committed.

No behaviour changes; the moves are pure renames and the import rewrites that follow them.
