+++ plugins/nemo-profiler/pyproject.toml
@@ -2,8 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 [project]
-name = "nemo-datasets-plugin"
-description = "Dataset profiler for NeMo Platform filesets."
+name = "nemo-profiler-plugin"
+description = "Fileset profiler for NeMo Platform. Ships the dataset profiler today; model profiling is the expected next kind."
 requires-python = ">=3.11,<3.15"
 dependencies = [
     "nemo-platform-plugin",
@@ -18,7 +18,7 @@ nemo-platform-plugin = { workspace = true }
 nemo-platform-sdk = { workspace = true }
 
 # Deliberately contributes no `nemo.cli` entry point. The profiler runs as a job task
-# (`python -m nemo_datasets_plugin.tasks.profile`), which is invoked by the platform rather than
+# (`python -m nemo_profiler_plugin.tasks.profile`), which is invoked by the platform rather than
 # typed by a user, so its inputs can keep moving while the feature is new.
 
 [build-system]
@@ -26,7 +26,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/nemo_datasets_plugin"]
+packages = ["src/nemo_profiler_plugin"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
