+++ plugins/nemo-profiler/src/nemo_profiler_plugin/dataset/readers/jsonl.py
@@ -9,8 +9,8 @@ import json
 from collections.abc import Iterator
 from typing import BinaryIO
 
-from nemo_datasets_plugin.profiler.file_source import FileEntry, FileSource
-from nemo_datasets_plugin.profiler.readers.base import FilePreview, ReadResult, register_reader
+from nemo_profiler_plugin.dataset.readers.base import FilePreview, ReadResult, register_reader
+from nemo_profiler_plugin.source import FileEntry, FileSource
 
 # Rows per batch, matching the parquet reader so the caller's working set does not depend on the
 # format being folded.
