+++ pytest.ini
@@ -10,7 +10,7 @@ pythonpath =
     .
     agents/nemo-studio-assistant/src
     plugins/example-plugin/src
-    plugins/nemo-datasets/src
+    plugins/nemo-profiler/src
     plugins/nemo-deployments/src
     plugins/nemo-deployments/tests/unit
     plugins/nemo-deployments/tests/integration
